pkg/controller: create namespace for svc if needed
This commit adds functionality to the controller to create and delete
namespaces for mirrored Services as necessary. It is entirely possible
that consumed remote Services will be deployed to a namespace that does
not exist locally. In this case, the controller must first create the
namespace before creating the Service and Endpoints. This commit also
allows the controller to delete a local namespace if it was reflected
and no Services are left in it.